refactor: clean up continue
